create base component for waiting room
